fix(theme): misaligned outline indicator
